+++ R/utils-prompt.R
@@ -1,7 +1,9 @@
-check_extension_approval <- function(
+check_approval <- function(
   no_prompt = FALSE,
   what = "Something",
-  see_more_at = NULL
+  see_more_at = NULL,
+  prompt_message = NULL,
+  interactive_info = NULL # could use `{ what }` as used in `cli_inform()`
 ) {
   if (no_prompt) return(TRUE)
 
@@ -14,22 +16,40 @@ check_extension_approval <- function(
       }
     ))
   } else {
-    cli::cli_inform(c(
-      "{what} may execute code when documents are rendered. ",
-      "*" = "If you do not trust the author(s) of this {what}, we recommend that you do not install or use this {what}."
-    ))
-    prompt_value <- tolower(readline(sprintf(
-      "Do you trust the authors of this %s (Y/n)? ",
-      what
-    )))
-    if (!prompt_value %in% "y") {
-      cli::cli_inform("{what} not installed.")
-      return(invisible(FALSE))
-    } else {
-      return(invisible(TRUE))
+    if (!is.null(interactive_info)) {
+      cli::cli_inform(interactive_info)
     }
-    return(invisible(TRUE))
+    prompt_value <- tolower(readline(prompt_message))
+    if (!prompt_value %in% c("", "y")) {
+      cli::cli_inform(paste0(what, " not installed."))
+    }
+    return(invisible(FALSE))
   }
+  return(invisible(TRUE))
+}
+
+check_extension_approval <- function(
+  no_prompt = FALSE,
+  what = "Something",
+  see_more_at = NULL
+) {
+  interactive_info <- c(
+    "{what} may execute code when documents are rendered. ",
+    "*" = "If you do not trust the author(s) of this {what}, we recommend that you do not install or use this {what}."
+  )
+
+  prompt_message <- sprintf(
+    "Do you trust the authors of this %s (Y/n)? ",
+    what
+  )
+
+  check_approval(
+    no_prompt = no_prompt,
+    what = what,
+    see_more_at = see_more_at,
+    prompt_message = prompt_message,
+    interactive_info = interactive_info
+  )
 }
 
 check_removal_approval <- function(
@@ -37,31 +57,16 @@ check_removal_approval <- function(
   what = "Something",
   see_more_at = NULL
 ) {
-  if (no_prompt) return(TRUE)
+  prompt_message <- sprintf(
+    "Are you sure you'd like to remove %s (Y/n)? ",
+    what
+  )
 
-  if (!is_interactive()) {
-    cli::cli_abort(c(
-      "{ what } requires explicit approval.",
-      ">" = "Set {.arg no_prompt = TRUE} if you agree.",
-      if (!is.null(see_more_at)) {
-        c(i = "See more at {.url {see_more_at}}")
-      }
-    ))
-  } else {
-    prompt_value <- tolower(readline(sprintf(
-      "? Are you sure you'd like to remove %s (Y/n)? ",
-      what
-    )))
-    if (!prompt_value %in% "y") {
-      return(invisible(FALSE))
-    }
-    return(invisible(TRUE))
-  }
-}
-
-# Add binding to base R function for testthat mocking
-readline <- NULL
-# Add binding to function from other package for mocking later on
-is_interactive <- function(...) {
-  rlang::is_interactive(...)
+  check_approval(
+    no_prompt = no_prompt,
+    what = what,
+    see_more_at = see_more_at,
+    prompt_message = prompt_message,
+    interactive_info = NULL
+  )
 }